# docs/swagger.yaml
basePath: /api
definitions:
...
  domain.CustomerStoreRequest:
    properties:
      email:
        type: string
      name:
        type: string
      phone_number:
        type: string
    required:
    - email
    - name
    - phone_number
    type: object
  domain.CustomerUpdateRequest:
    properties:
      email:
        type: string
      name:
        type: string
      phone_number:
        type: string
    type: object
  domain.Error:
    properties:
      code:
        type: integer
      errors:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
  domain.Success:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  title: Book Store API Documentation
  version: "1.0"
paths:
...
  /customers:
    get:
      consumes:
      - application/json
      description: Get list of customers
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Size of page (default 10)
        in: query
        name: size
        type: integer
      - description: Search query
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of customers
          schema:
            items:
              $ref: '#/definitions/domain.Success'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Get list of customer
      tags:
      - customers
    post:
      consumes:
      - application/json
      description: Store customer
      parameters:
      - description: customer data
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/domain.CustomerStoreRequest'
      produces:
      - application/json
      responses:
        "201":
          description: customer detail
          schema:
            $ref: '#/definitions/domain.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      security:
      - Bearer: []
      summary: Store customer
      tags:
      - customers
  /customers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete customer
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success delete customer
          schema:
            $ref: '#/definitions/domain.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      security:
      - Bearer: []
      summary: Delete customer
      tags:
      - customers
    get:
      consumes:
      - application/json
      description: Get customer by id
      parameters:
      - description: customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: customer detail
          schema:
            $ref: '#/definitions/domain.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Get customer by id
      tags:
      - customers
    put:
      consumes:
      - application/json
      description: Update customer
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Customer data
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/domain.CustomerUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Customer detail
          schema:
            $ref: '#/definitions/domain.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      security:
      - Bearer: []
      summary: Update customer
      tags:
      - customers
  /roles:
    get:
      consumes:
      - application/json
      description: Get list of roles
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Size of page (default 10)
        in: query
        name: size
        type: integer
      - description: Search query
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of roles
          schema:
            items:
              $ref: '#/definitions/domain.Success'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Get list of role
      tags:
      - roles
  /roles/{id}:
    get:
      consumes:
      - application/json
      description: Get role by id
      parameters:
      - description: role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: role detail
          schema:
            $ref: '#/definitions/domain.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Get role by id
      tags:
      - roles
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"